+++ packages/staker/src/matic/utils.js
@@ -7,7 +7,6 @@ import {jsonRpcFetch} from '../helpers/jsonRpc';
 
 let cachedMaticNodeUrl = '';
 
-
 /**
  * Random pick a MATIC node.
  *
@@ -21,7 +20,3 @@ let cachedMaticNodeUrl = '';
 
 export const maticFetch = (fetch, body) =>
   jsonRpcFetch(fetch, getMaticNodeURL(), body);
-
-// self deployed ethscan contract
-// https://github.com/MyCryptoHQ/eth-scan/blob/master/contracts/BalanceScanner.sol
-export const MATIC_ETHSCAN_CONTRACT = '0xd570a3C50D157A58625e33C69F1542a5e5594b46';